// src/components/card/model/cardTypes.ts
export interface PlacesType {
    name: string;
    id: number;
    rating: number;
    address: string;
    cuisine: string;
    price_range: string;
    wait_time: number;
    avg_price_range: string;
    seats: number;
    wifi: boolean;
    music: boolean;
    kids_friendly: string;
    parking: string;
    working_hours: string;
    image_url: string;
}

export interface RestaurantType {
    name: string;
    id: number;
    rating: number;
    address: string;
    cuisine: string;
    price_range: string;
    wait_time: number;
    avg_price_range: string;
    seats: number;
    wifi: boolean;
    music: boolean;
    kids_friendly: string;
    parking: string;
    working_hours: string;
    image_url: string;
}